+++ adjective.js
@@ -1,0 +1,11 @@
+function Adjective() {
+  this.happy = true;
+  this.sad = true;
+  this.mad = true;
+  this.orange = true;
+  this.pink = true;
+  this.rotten = true;
+  this.bad = true;
+};
+
+module.exports.theAdjective = Adjective;
