+++ app/index.html
@@ -3,11 +3,17 @@
 <body>
   <h1>Random Band Name Generator</h1>
 
-        <button id='name'>click me for random band name.</button>
-        <p>
-          name:
-          <span id='adjective'></span> <span id='verb'></span> <span id='noun'></span>
-        </p>
+  <button id='name'>click me for random band name</button>
+    <p>
+      name:
+      <span id='adjective'></span> <span id='verb'></span> <span id='noun'></span>
+    </p>
+  <h3>Add an adjective to the list</h3>
+    <form action="http://localhost:3000/adjective" method="POST" id"newAdjective">
+     <input type="text"  placeholder="TYPE ADJECTIVE HERE" id="adject">
+     <button id='submitAdj'>add adjective to list</button>
+    </form>
+
 </body>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
 <script src="./script.js"></script>
