+++ app/script.js
@@ -3,19 +3,28 @@ $(function () {
   var $adj = $('#adjective');
   var $noun = $('#noun');
   var $verb = $('#verb');
+  var $submitAdj = $('#submitAdj');
+  var $adject = $('#adject');
 
   $button.click(function () {
-    $.post('http://localhost:3000/adjective', function (response) {
+    $.get('http://localhost:3000/adjective', function (response) {
       $adj.text(response.word);
     });
 
-    $.post('http://localhost:3000/verb', function (response) {
+    $.get('http://localhost:3000/verb', function (response) {
       $verb.text(response.word);
     });
 
-    $.post('http://localhost:3000/noun', function (response) {
+    $.get('http://localhost:3000/noun', function (response) {
       $noun.text(response.word);
     });
   });
+
+  $submitAdj.click(function () {
+
+  });
+
 });
+
+
