+++ noun.js
@@ -1,0 +1,10 @@
+function Noun() {
+  this.squirrel = true;
+  this.weinerDogs = true;
+  this.fingerNail = true;
+  this.eggMcMuffin = true;
+  this.sneakers = true;
+  this.medicine = true;
+};
+
+module.exports.theNoun = Noun;
