+++ verb.js
@@ -1,0 +1,10 @@
+function Verb() {
+  this.running = true;
+  this.swimming = true;
+  this.canoeing = true;
+  this.gambling = true;
+  this.driving = true;
+  this.smelling = true;
+};
+
+module.exports.theVerb = Verb;
